Move file lock functions to their own module.
